Update .env
this uses two methods to buy and has a dynamic profit capture function that ensure a small profit is captured if the price goes halfway to the tp value. this could make a few profitable trades close early but there is no limit on how often it buys so it will reenter almost immediately.