Uppated Binance.py
This one has two buy triggers, using 3 pieces of logic. 
If atr breakout is big and trend is going up, if ema crossover and if sma percent difference is large enough. There are OR'd so any of them can trigger. 
Static TP is set to twice the distance as the SL is from entry price and the SL follows the price up to the TP to maximise profits. 
Lots of tweaking needs to be done to get this working appropriately and also it needs to be put onto a cloud server as my phone keeps changing IPs and killing the background processes.